Moved all stories into their own folder
